Add methods to interrogate the queue
To know if a job exists for a task, if it's currently active and get the information about the next job for a task.

I needed this to know if a job was currently active for a screen so I could adapt the UI to tell the user that it's still processing. My tasks have unique names that I can recompose to find them in the queue.